# Xdoc: Add a hack to preprocessing to make save/render faster.