This change speeds up the manual build (certification of doc/top.lisp) by around 3-5 minutes by adding a semi-ugly hack to xdoc's get-event, which it uses for expanding preprocessor directives like @(def ...) and @(thm ...).  The idea is that before we start, we traverse the world once and collect a fast alist binding every function/theorem/macro name to the event form we'd get from get-event*.  (Previously, each call of get-event would traverse the world linearly, which took a lot of time -- around 90 seconds in save-rendered with doc/top loaded.) This fast alist is generated and bound in a state global at the beginning of the save/render routine.  If it isn't bound, everything works as before, which is important for using :doc/:xdoc from the terminal.

Submitting this as a pull request in case anyone wants to weigh in -- @MattKaufmann @ragerdl @jaredcdavis