@MattKaufmann 2017-11-10
Looks like a cool idea to me!  I totally trust that you did it right (and besides, if not then we'll probably find out ;).  If nobody else comments soon, please feel free to accept your own pull request, as far as I'm concerned (or I will).