@jaredcdavis 2017-11-10
Hi,

Sounds awesome.  Thanks!

Cheers,
Jared

On Fri, Nov 10, 2017 at 4:29 PM, MattKaufmann <notifications@github.com> wrote:
>
> Looks like a cool idea to me! I totally trust that you did it right (and besides, if not then we'll probably find out ;). If nobody else comments soon, please feel free to accept your own pull request, as far as I'm concerned (or I will).
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.




-- 
Jared C. Davis <jared.c.davis@gmail.com>
11410 Windermere Meadows
Austin, TX 78759
http://www.cs.utexas.edu/users/jared/