@ragerdl 2017-11-11
Super-excited you found a reasonable way to speed-up our build bottleneck!  Merging, since everyone's chimed in.