@MattKaufmann 2017-11-11
Hi, Sol --

I'll have to do some runs to figure that out.  I'll do so and report
back when that's done.

-- Matt
solswords <notifications@github.com> writes:

> [1:text/plain Hide]
>
> Hi Matt,
> Just curious - how long does CCL take to certify that book on your system?
> -Sol
>
> On Nov 11, 2017 3:03 PM, "MattKaufmann" <notifications@github.com> wrote:
>
>> By the way, I ran an experiment, building the manual using
>> an SBCL-based ACL2 executable, both before and after
>> updating. The difference was noticeable but small, as shown
>> below. I can easily imagine that CCL would produce a bigger
>> difference because of static consing.
>>
>> ;;; old
>> Built [...]/books/doc/top.cert (489.394s)
>> 443.548u 14.716s 8:24.49 90.8%  0+0k 2736+531360io 1pf+0w
>>
>> ;;; new
>> Built [...]/books/doc/top.cert (473.852s)
>> 428.656u 14.960s 8:08.77 90.7%  0+0k 5960+531344io 0pf+0w
>>
>> —
>> You are receiving this because you authored the thread.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/acl2/acl2/pull/795#issuecomment-343694015>, or mute
>> the thread
>> <https://github.com/notifications/unsubscribe-auth/AFKzKSpbrW8WLbbHJQ_Wpny_4NYnW5_Rks5s1guNgaJpZM4QaIAA>
>> .
>>