@MattKaufmann 2017-11-12
Hi, Sol --

Here are results with CCL.  The "git pull" I did updated from
something pretty recent, so we can reasonably attribute all of the
time difference to your change.

;;; old
Built /v/filer4b/v11q002/acl2space/acl2/acl2/books/doc/top.cert (979.578s)
921.244u 8.436s 16:56.12 91.4%	0+0k 1307424+531824io 219pf+0w

;;; new
Built /v/filer4b/v11q002/acl2space/acl2/acl2/books/doc/top.cert (841.411s)
801.364u 7.752s 14:15.73 94.5%	0+0k 7960+532016io 0pf+0w

Not bad!

A possible explanation for why there isn't more speed-up is that
get-event isn't terribly slow -- there is some world indexing that's
supposed to help it, I think.  I haven't thought this through or even
absorbed what you did, so please ignore this remark if it doesn't make
much sense.

-- Matt
Matt Kaufmann <kaufmann@cs.utexas.edu> writes:

> Hi, Sol --
>
> I'll have to do some runs to figure that out.  I'll do so and report
> back when that's done.
>
> -- Matt
> solswords <notifications@github.com> writes:
>
>> [1:text/plain Hide]
>>
>> Hi Matt,
>> Just curious - how long does CCL take to certify that book on your system?
>> -Sol
>>
>> On Nov 11, 2017 3:03 PM, "MattKaufmann" <notifications@github.com> wrote:
>>
>>> By the way, I ran an experiment, building the manual using
>>> an SBCL-based ACL2 executable, both before and after
>>> updating. The difference was noticeable but small, as shown
>>> below. I can easily imagine that CCL would produce a bigger
>>> difference because of static consing.
>>>
>>> ;;; old
>>> Built [...]/books/doc/top.cert (489.394s)
>>> 443.548u 14.716s 8:24.49 90.8%  0+0k 2736+531360io 1pf+0w
>>>
>>> ;;; new
>>> Built [...]/books/doc/top.cert (473.852s)
>>> 428.656u 14.960s 8:08.77 90.7%  0+0k 5960+531344io 0pf+0w
>>>
>>> —
>>> You are receiving this because you authored the thread.
>>> Reply to this email directly, view it on GitHub
>>> <https://github.com/acl2/acl2/pull/795#issuecomment-343694015>, or mute
>>> the thread
>>> <https://github.com/notifications/unsubscribe-auth/AFKzKSpbrW8WLbbHJQ_Wpny_4NYnW5_Rks5s1guNgaJpZM4QaIAA>
>>> .
>>>